Feat: split login page and loading page
