+++ src/pages/GoogleLogin.tsx
@@ -2,9 +2,9 @@ import { useMutation } from '@tanstack/react-query';
 import axios from 'axios';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { PacmanLoader } from 'react-spinners';
+import Loading from './Loading';
 
-const Loading = () => {
+const GoogleLogin = () => {
   const navigate = useNavigate();
 
   // 게시판 페이지로 이동
@@ -45,14 +45,7 @@ const Loading = () => {
     }
   }, [code]);
 
-  return (
-    <div className="flex flex-col items-center justify-center h-screen w-screen gap-y-4">
-      <div className="font-pretendard font-bold text-black ">
-        잠시만 기다려주세요
-      </div>
-      <PacmanLoader color="#FFC939" />
-    </div>
-  );
+  return <Loading />;
 };
 
-export default Loading;
+export default GoogleLogin;
